demux: move metadata-based replaygain decoding out of af_volume
